Refactor Move class (#42)
Move Move into move subpackage in preparation for refactor.
