+++ src/main/java/leelawatcher/gui/BoardView.java
@@ -24,7 +24,7 @@ package leelawatcher.gui;
 
 import leelawatcher.goboard.Board;
 import leelawatcher.goboard.IllegalMoveException;
-import leelawatcher.goboard.Move;
+import leelawatcher.goboard.move.Move;
 import leelawatcher.goboard.PointOfPlay;
 
 import java.awt.*;
